// packages/backend/src/server/api/stream/types.ts
export type NowPlayingSong = {
    coverArtUrl: string;
    title: string;
    artist: string;
    album: string;
    url: string;
    progressMs: number;
    durationMs: number;
    isPlaying: boolean;
    service: 'spotify' | 'lastfm';
};